refactor: package 분리
- model 내부 package : lotto, rank 로 분리
