+++ src/main/java/lotto/model/lotto/LottoNumber.java
@@ -1,4 +1,4 @@
-package lotto.model;
+package lotto.model.lotto;
 
 import lotto.exception.InvalidLottoNumberException;
