+++ src/main/java/lotto/model/lotto/Money.java
@@ -1,4 +1,4 @@
-package lotto.model;
+package lotto.model.lotto;
 
 import java.util.Objects;
