+++ src/main/java/lotto/model/rank/Ranks.java
@@ -1,4 +1,4 @@
-package lotto.model;
+package lotto.model.rank;
 
 import java.util.Collections;
 import java.util.LinkedHashMap;
