+++ src/test/java/lotto/model/LottoNumberTest.java
@@ -1,6 +1,7 @@
 package lotto.model;
 
 import lotto.exception.InvalidLottoNumberException;
+import lotto.model.lotto.LottoNumber;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
