+++ src/test/java/lotto/model/RankResultsTest.java
@@ -1,5 +1,9 @@
 package lotto.model;
 
+import lotto.model.lotto.Money;
+import lotto.model.rank.Rank;
+import lotto.model.rank.RankResult;
+import lotto.model.rank.RankResults;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
