+++ src/test/java/lotto/model/WinningTicketTest.java
@@ -1,6 +1,10 @@
 package lotto.model;
 
 import lotto.exception.InvalidLottoTicketException;
+import lotto.model.lotto.LottoNumber;
+import lotto.model.lotto.LottoTicket;
+import lotto.model.lotto.WinningTicket;
+import lotto.model.rank.Rank;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
